+++ ethers-middleware/src/gas_oracle/provider_oracle.rs
@@ -1,4 +1,4 @@
-use crate::gas_oracle::{GasOracle, GasOracleError};
+use super::{GasOracle, GasOracleError, Result};
 use async_trait::async_trait;
 use ethers_core::types::U256;
 use ethers_providers::Middleware;
@@ -6,7 +6,8 @@ use std::fmt::Debug;
 
 /// Gas oracle from a [`Middleware`] implementation such as an
 /// Ethereum RPC provider.
-#[derive(Debug)]
+#[derive(Clone, Debug)]
+#[must_use]
 pub struct ProviderOracle<M: Middleware> {
     provider: M,
 }
@@ -21,16 +22,16 @@ impl<M: Middleware> ProviderOracle<M> {
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 impl<M: Middleware> GasOracle for ProviderOracle<M>
 where
-    <M as Middleware>::Error: 'static,
+    M::Error: 'static,
 {
-    async fn fetch(&self) -> Result<U256, GasOracleError> {
+    async fn fetch(&self) -> Result<U256> {
         self.provider
             .get_gas_price()
             .await
             .map_err(|err| GasOracleError::ProviderError(Box::new(err)))
     }
 
-    async fn estimate_eip1559_fees(&self) -> Result<(U256, U256), GasOracleError> {
+    async fn estimate_eip1559_fees(&self) -> Result<(U256, U256)> {
         // TODO: Allow configuring different estimation functions.
         self.provider
             .estimate_eip1559_fees(None)